Dockerizing the go app
