+++ .env
@@ -22,12 +22,12 @@ TEST_REDIS_NAMESPACE=test
 TEST_INSTACHAT_REDIS_DB=1
 TEST_CONFIGURATIONS_CACHE_REDIS_DB=13
 
-DEVELOPMENT_REDIS_URL_GO_APP=localhost:6379
+DEVELOPMENT_REDIS_URL_GO_APP=instachat_redis:6379
 
 DEVELOPMENT_ASSET_HOST=http://localhost:3000
 TEST_ASSET_HOST=http://example.com
 # Mailer Host
-DEVELOPMENT_HOST=http://localhost:3000
+DEVELOPMENT_HOST=http://instachat:3000
 TEST_HOST=http://example.com
 STAGING_HOST=
 PRODUCTION_HOST=
