--- dotenv_load.go
@@ -1,13 +1,0 @@
-package main
-
-import (
-    "github.com/joho/godotenv"
-)
-
-func init() {
-
-    err := godotenv.Load(".env")
-		if err != nil { panic(err.Error()) }
-
-}
-
