--- go.mod
@@ -1,9 +1,0 @@
-module instchat_endpoints
-
-go 1.15
-
-require (
-	github.com/garyburd/redigo v1.6.2
-	github.com/go-sql-driver/mysql v1.5.0
-	github.com/joho/godotenv v1.3.0
-)
