+++ go_app/worker.go
@@ -1,0 +1,5 @@
+package main
+
+type Worker interface {
+	Perform()
+}
