--- worker.go
@@ -1,5 +1,0 @@
-package main
-
-type Worker interface {
-	Perform()
-}
